+++ src/app/admin/Components/user-list/user-list.component.ts
@@ -1,6 +1,10 @@
 import { Component } from '@angular/core';
 import { User } from '../../../user/Models/user';
 import { UserService } from '../../../user/Services/user.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../../Dialog/confirmation-dialog/confirmation-dialog.component';
+import { SnackbarService } from '../../../user/Services/snackbar.service';
+import { error } from 'console';
 
 @Component({
   selector: 'app-user-list',
@@ -8,7 +12,7 @@ import { UserService } from '../../../user/Services/user.service';
   styleUrl: './user-list.component.scss'
 })
 export class UserListComponent {
-  constructor(private userService:UserService) {}
+  constructor(private userService:UserService,private dialog:MatDialog,private snackBar:SnackbarService) {}
 
 
   users:User[]=[];
@@ -23,11 +27,25 @@ export class UserListComponent {
   }
 deleteUser(id:string){
   console.log("delete");
+  const dialogRef=this.dialog.open(ConfirmationDialogComponent,{width:"400px",height:"auto"});
+  dialogRef.afterClosed().subscribe((accept)=>{
+    if(!accept)
+    {
+      return false;
+    }
+    else
+    {
+      this.userService.deleteUser(id).subscribe((data:any)=>{
+        console.log(data);
+        this.snackBar.snackbarnotification("User Deleted Successfully","Dismiss",'success');
+        this.getUserList();
+      })
+      return true;
+    }
+  },(error=>{
+    this.snackBar.snackbarnotification("Something Went Wrong","Dismiss",'error');
+  }))
   
-  this.userService.deleteUser(id).subscribe((data:any)=>{
-    console.log(data);
-    this.getUserList();
-  })
 }
 ngOnInit():void{
   this.getUserList();
